requirements.txt
streamlit>=1.30.0
openai>=1.12.0
python-docx>=1.1.0
reportlab>=4.1.0
